.github/workflows: fix CI.

+++ surfify/tests/test_models_unet.py
@@ -89,7 +89,7 @@ class TestModelsGUNet(unittest.TestCase):
             for ch_idx in range(X.shape[1]):
                 _X.append(utils.text2grid(ico_vertices, X[sample_idx, ch_idx]))
             self.X.append(_X)
-        self.X = np.asarray(self.X)
+        self.X = np.asarray(self.X).astype(np.float32)
         self.y = utils.text2grid(ico_vertices, y)
         self.X = torch.from_numpy(self.X)
         self.y = torch.from_numpy(self.y)
